backend/Yiqi/Yiqi/requirements.txt
appnope==0.1.0
argh==0.26.2
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.24.0
attrs==17.3.0
autobahn==18.3.1
Automat==0.6.0
beautifulsoup4==4.6.0
bleach==2.1.3
bs4==0.0.1
certifi==2018.1.18
cffi==1.11.2
channels==1.1.8
chardet==3.0.4
click==6.7
cloudinary==1.11.0
configparser==3.5.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4.2
crypto==1.4.1
cryptography==2.1.4
cssselect==1.0.1
cycler==0.10.0
daphne==1.4.2
decorator==4.2.1
defusedxml==0.5.0
diff-match-patch==20121119
dj-database-url==0.5.0
Django==1.11.7
django-autoslug==1.9.3
django-cors-headers==2.1.0
django-crispy-forms==1.7.0
django-extensions==2.0.6
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==0.6.1
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-redis==4.9.0
django-reversion==2.0.12
django-simple-captcha==0.5.6
django-taggit==0.22.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
docopt==0.6.2
docutils==0.14
drf-extensions==0.3.1
et-xmlfile==1.0.1
feedparser==5.2.1
Flask==0.12.2
future==0.16.0
html5lib==1.0.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
itsdangerous==0.24
itypes==1.1.0
jdcal==1.3
jedi==0.11.1
Jinja2==2.10
lockfile==0.12.2
luigi==2.7.1
lxml==4.1.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.2
mock==2.0.0
msgpack-python==0.5.6
mysqlclient==1.3.12
Naked==0.1.31
np==1.0.2
oauthlib==2.0.7
odfpy==1.3.6
olefile==0.44
openpyxl==2.4.9
optionaldict==0.1.1
parsel==1.2.0
parso==0.1.1
pathtools==0.1.2
pbr==3.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==4.3.0
pluggy==0.6.0
prettytable==0.7.2
project==20
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
py==1.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.6.0
PyDispatcher==2.0.5
pygame==1.9.3
Pygments==2.2.0
PyJWT==1.6.1
PyMySQL==0.8.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
pytesseract==0.1.7
pytest==3.3.2
python-daemon==2.1.2
python-dateutil==2.6.1
python-decouple==3.1
python-memcached==1.59
python3-openid==3.1.0
pytz==2017.3
PyYAML==3.12
queuelib==1.4.2
redis==2.10.6
reload==0.9
requests==2.18.4
requests-oauthlib==0.8.0
rncryptor==3.2.0
Scrapy==1.4.0
service-identity==17.0.0
shellescape==3.4.1
simplegeneric==0.8.1
six==1.11.0
sleekxmpp==1.3.3
social-auth-app-django==2.1.0
social-auth-core==1.7.0
stevedore==1.28.0
tablib==0.12.1
tornado==4.5.3
tox==2.9.1
traitlets==4.3.2
Twisted==17.9.0
txaio==2.9.0
unicodecsv==0.14.1
Unipath==1.1
uritemplate==3.0.0
urllib3==1.22
uWSGI==2.0.15
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
w3lib==1.18.0
watchdog==0.8.3
wcwidth==0.1.7
web.py==0.40.dev0
webencodings==0.5.1
wechatpy==1.6.0
Werkzeug==0.13
xlrd==1.1.0
xlwt==1.3.0
xmltodict==0.11.0
zope.interface==4.4.3
